Modify 3DES-CBC implementation to work with TLS 1.0/1.1 (e.g. DES-CBC3-SHA)
The design of our 3DES-CBC implementation up to this point is tied to the notion
of "update/final." That is, it expects that N bytes of data will be fed to it to
encrypt/decrypt via 1 or more calls to update (e.g. EVP_EncryptUpdate /
EVP_DecryptUpdate) followed by a call to final (e.g.
EVP_EncryptFinal_ex), which encrypts/decrypts any remaining data in the last
block and adds padding. This appears to be a fine design choice until we come to
a TLS 1.0 function in OpenSSL, tls1_enc. This function calls the deprecated
function EVP_Cipher (documentation instructs users to use EVP_CipherUpdate and
EVP_CipherFinal_ex instead) and adds the padding itself. This doesn't work with
our design because we'll be forever waiting for a call to final to finish the
last block of data and add padding. As far as I could tell while debugging the
code, there's no way for our current implementation to detect this scenario and
encrypt/decrypt the whole payload in one shot, without adding padding. So, our
options are then 1) Modify OpenSSL to work with the existing design
(not preferable) and 2) remove the EVP_CIPH_FLAG_CUSTOM_CIPHER flag and let
OpenSSL handle padding and the last bit of data for us. 2) also requires that we
simplify the encrypt/decrypt functions to just encrypt/decrypt the buffer
they're given and not have any notion of update/final.