Fix issue with AES-GCM decryption benchmarking.
We are ignoring an error from `EVP_DecryptFinal_ex` since we expect the tag to
not match the data. I think we also need to ignore the same type of error on the
call to `EVP_DecryptUpdate`. Currently, I'm seeing `AES_GCM_AUTH_E` being
returned by wolfCrypt, which causes the benchmark to fail.