Add support for update/final-style AES-GCM to we_aes_gcm_cipher.
OpenSSL's AES-GCM code allows you to call EVP_CipherUpdate() with just the
ciphertext and no tag. Then, you set the tag and call EVP_CipherFinal() to
check the tag. For the FIPS v2 wolfCrypt module, there's no way to do an
AES-GCM decryption without checking the tag. To work around this, defining
WE_AES_GCM_DECRYPT_ON_FINAL will do the following:

- If we_aes_gcm_cipher is called for decryption when the tag hasn't been set,
the ciphertext is saved in a temporary buffer on the we_AesGcm object.
- Then, when we_aes_gcm_cipher is called by EVP_CipherFinal() (or similar
"final" function), we decrypt the saved ciphertext and check the tag.

WE_AES_GCM_DECRYPT_ON_FINAL is not defined by default. Defining it fixes an
error when using wolfEngine with OpenVPN and OpenSSL 1.0.2 or later.