Fix bug with AES-CTR initialization.
OpenSSL allows the user to call EVP_CipherInit with NULL key or IV. Prior to
this commit, setting the IV first (with key NULL) and then setting the key (with
IV NULL) would result in the IV getting set to 0s on the call to set the key.
This was discovered in testing with OpenSSH. This commit fixes this bug and adds
a regression test for it.