Fix ECDSA verify error return code.
From the OpenSSL docs: "ECDSA_verify() and ECDSA_do_verify() return 1 for a
valid signature, 0 for an invalid signature and -1 on error." Our verify
functions are returning 0 for both invalid signature (i.e. verify failed) and
other errors. This commit makes it so all errors that aren't verification
failure cause a -1 return. This was discovered by running the OpenSSL 1.1.1m
unit tests with wolfEngine.